Update
HUD is started: will need the txt placed underneath them to be on top of
the GUIs created as well as a script to change the text information.

I have added some street lights with the intent to add different color
glows in order to make the maze areas unique and not so blah as Patrick
had put it. I think adding some gaffiti to the walls would help with
this as well. There are also some decor like park benches to once again
add to the environment.