Worked on the lighting.
I made a second copy of the project because I did not want to overwrite
your existing modifications.  However, due to the large quantity of
lights in the scene, the game was not rendering properly.  It was for
that reason I removed all the lights and made changes to them.  Take a
look and let me know what you think.  Ignore the fact that I temporarily
removed the other models.  I have to adjust them in Maya.