Transition btw Scenes and Fixed Most Models
We now have a start scene, a game scene, and a credits scene that
transition between each other.  I fixed the benches and wood crate
models.  Their center points are now on the bottom and centered so
placing them will be easier.  I also created prefabs for them with
colliders to be used throughout the scene.  The rest of the models will
be fixed later.