Battery/PowerBar GUI and Player Texture for MiniMap
I have created the battery texture and also a powerbar texture and
scripted them into the Hud Script.
The power bar will need instructions to Shrink its width over the time
period that the flashlight remains active.
The powerbar texture (originally white) should start as a nice green
color and once it hits 75% of the original width to switch to a
light green color. 50% of the original width will change the powerbar
texture to Yellow and 25% of the original width to Red.
I have placed a big white flat sphere shape over the playerobject so
that it is very visible to the player in the minimap.
As for the minimap, I struggled to get it to adjust to different screen
resolution, but I have it attached to the game controller.
I am working on the credit list right now... and will post it later or
tomorrow morning depending on my fiance's schedule.