Few of the issues from the to-do list are complete.
Unfortunately, I did not get to the dialog tonight, but it will be one
of the first things I do on Tuesday when I get off work.  Flashlight
meter now drains, but flashlight remains on.  Life count decreases, but
game only removes player's control when number of lives is zero.
Basically, the only thing left is to do the coding, which may take some
time.  Hopefully, I will be able to get off work early to get it all
done before midnight.