Whatnots
For the Credit Script I had added information to it and readjusted it
the best I could.
I also removed the 'press spacebar or enter to continue' Gui from Credit
Screen. It is not needed if it goes right to the main menu.
I readjusted the battery so it didn't take up so much of the screen.
I added another light source just to brighten it up just a little bit.
I added the Spacebar and Enter functions to the MenuScreen Script.